switch endpoint from /material-inventory -> /materials/inventory
