+++ application/controllers/materialController.js
@@ -4,15 +4,21 @@ const { verifyJwtToken } = require('../middleware/authorize');
 const VendorModel = require('../models/vendor');
 const MaterialCategoryModel = require('../models/materialCategory');
 
+const materialInventoryService = require('../services/materialInventoryService');
+const materialService = require('../services/materialService');
+const purchaseOrderService = require('../services/purchaseOrderService');
+const ticketService = require('../services/ticketService');
+const mongooseService = require('../services/mongooseService');
+
 const SHOW_ALL_MATERIALS_ENDPOINT = '/materials';
 
 router.use(verifyJwtToken);
 
 router.get('/all', async (request, response) => {
     try {
-        
+
         const materials = await MaterialModel.find().exec();
-        
+
         return response.send(materials);
 
     } catch (error) {
@@ -24,7 +30,7 @@ router.get('/all', async (request, response) => {
 router.get('/', async (request, response) => {
     try {
         const materials = await MaterialModel.find().exec();
-        
+
         return response.render('viewMaterials', {
             materials: materials
         });
@@ -62,9 +68,9 @@ router.get('/update/:id', async (request, response) => {
         const materialCategories = await MaterialCategoryModel.find().exec();
 
         return response.render('updateMaterial', {
-            material, 
+            material,
             vendors,
-            materialCategories 
+            materialCategories
         });
     } catch (error) {
         console.log(error);
@@ -99,6 +105,49 @@ router.get('/delete/:id', async (request, response) => {
     }
 
     return response.redirect('back');
+});
+
+router.get('/inventory', async (request, response) => {
+    console.log('hiiii');
+    try {
+        const allMaterials = await materialService.getAllMaterials();
+
+        const distinctMaterialObjectIds = mongooseService.getObjectIds(allMaterials);
+        const distinctMaterialIds = materialService.getMaterialIds(allMaterials);
+
+        const allPurchaseOrders = await purchaseOrderService.getPurchaseOrdersForMaterials(distinctMaterialObjectIds);
+        const materialIdToPurchaseOrders = materialInventoryService.mapMaterialIdToPurchaseOrders(distinctMaterialObjectIds, allPurchaseOrders);
+
+        const purchaseOrdersThatHaveArrived = purchaseOrderService.findPurchaseOrdersThatHaveArrived(allPurchaseOrders);
+        const purchaseOrdersThatHaveNotArrived = purchaseOrderService.findPurchaseOrdersThatHaveNotArrived(allPurchaseOrders);
+
+        const lengthOfAllMaterialsInInventory = purchaseOrderService.computeLengthOfMaterial(purchaseOrdersThatHaveArrived);
+        const lengthOfAllMaterialsOrdered = purchaseOrderService.computeLengthOfMaterial(purchaseOrdersThatHaveNotArrived);
+
+        const materialObjectIdToLengthUsedByTickets = await ticketService.getLengthOfEachMaterialUsedByTickets(distinctMaterialIds);
+
+        const materialInventories = allMaterials.map((material) => {
+            const feetOfMaterialAlreadyUsedByTickets = materialObjectIdToLengthUsedByTickets[material.materialId] || 0;
+            const purchaseOrdersForMaterial = materialIdToPurchaseOrders[material._id];
+
+            return materialInventoryService.buildMaterialInventory(material, purchaseOrdersForMaterial, feetOfMaterialAlreadyUsedByTickets);
+        });
+
+        const netLengthOfMaterialInInventory = materialInventoryService.computeNetLengthOfMaterialInInventory(materialInventories);
+
+        return response.send({
+            materialInventories,
+            lengthOfAllMaterialsInInventory,
+            lengthOfAllMaterialsOrdered,
+            netLengthOfMaterialInInventory
+        });
+
+    } catch (error) {
+        console.log(`An error occurred while attempting to load /material/inventory: ${error}`);
+
+        request.flash('errors', [`The following error occurred: ${error}`]);
+        return response.redirect('back');
+    }
 });
 
 module.exports = router;
