+++ application/views/viewMaterialInventory.ejs
@@ -22,10 +22,6 @@
         <span>Feet On Order</span>
         <h1 class='total-length-of-material-ordered'><%= lengthOfAllMaterialsOrdered || 'N/A' %></h1>
       </div>
-      <!-- <div class='card col col-five'>
-        <span>Total Open POs</span>
-        <h1 class='total-purchase-ordered'><%= totalPurchaseOrders || 'N/A' %></h1>
-      </div> -->
     </div>
 
     <div class="workflow-filter flex-center-left-row full-width card">
